admin: wire Categories/Items/Units/Markets to real API data (Epic 3 #31)
Categories.razor, Units.razor, Items.razor, MarketList.razor and
MarketDetail.razor (including delivery windows and the service-area
map) now call AdminApiClient instead of PlaceholderData, with full
CRUD (create/edit/delete) - the plan's Epic 3 deliverable is "Admin
CRUD", not read-only lists. Api/CatalogDtos.cs and Api/MarketDtos.cs
are Admin's own copy of the wire shapes (mirrors CatalogContracts.cs/
MarketContracts.cs without referencing the Application layer, same
pattern as AdminApiDtos.cs from #29). AdminViewModels' CategoryRow,
ItemRow, UnitRow, MarketRow and DeliveryWindowRow ids widened int to
long to match every id in the real API; PlaceholderData.cs itself is
untouched since DeliveryRuns.razor still reads PlaceholderData.Markets
and isn't part of this task.

Two scope calls worth recording:
- Item image upload is out of scope here - CreateItemRequest always
  sends a null MediaAssetId, and editing an item preserves whatever
  MediaAssetId it already had. Wiring a real upload flow (presign,
  direct-to-S3 PUT, commit) is a separable piece of work.
- MarketDetail's "Verified stalls" figure and MarketList's column of
  the same name are removed rather than wired to a fake 0 - no
  partner/stall verification system exists yet (a later epic), so
  there is nothing real to show.

Verified end-to-end in a real Chromium browser against the live API
and DB, covering the actual Epic 3 exit criterion: create a category,
unit and item; create a market; draw a real polygon with Leaflet.draw
and save it; edit the market to Active (rejected until the area was
saved, as the API requires); add/edit/remove a delivery window; then,
via the real customer-facing API, confirm a point inside the
admin-drawn polygon is accepted and a point outside is refused with a
clear message. Two false negatives during verification (a "close by
re-clicking the first vertex" gesture not registering, and a
combined-script timing race) were both testing-technique bugs,
confirmed by re-running the same steps in isolation - not application
bugs; no application code changed as a result.

+++ admin/src/OnlinePalengke.Admin/ViewModels/AdminViewModels.cs
@@ -35,7 +35,7 @@ public sealed record RiderOnShift(
 
 /// <summary>Catalog category. Categories group items and scope the per-category winning bid.</summary>
 public sealed record CategoryRow(
-    int Id,
+    long Id,
     string Name,
     string Slug,
     int ItemCount,
@@ -47,7 +47,7 @@ public sealed record CategoryRow(
 /// exist on partner quotes, so no screen in this app may show a price against an item.
 /// </summary>
 public sealed record ItemRow(
-    int Id,
+    long Id,
     string Name,
     string CategoryName,
     string DefaultUnitCode,
@@ -57,7 +57,7 @@ public sealed record ItemRow(
 
 /// <summary>Unit of measure available to shopping lists and quotes.</summary>
 public sealed record UnitRow(
-    int Id,
+    long Id,
     string Code,
     string Name,
     bool AllowsFractionalQuantity,
@@ -65,7 +65,7 @@ public sealed record UnitRow(
 
 /// <summary>A wet market in the network.</summary>
 public sealed record MarketRow(
-    int Id,
+    long Id,
     string Name,
     string City,
     string Province,
@@ -76,7 +76,7 @@ public sealed record MarketRow(
 
 /// <summary>A recurring delivery window for one market, expressed in Manila wall-clock time.</summary>
 public sealed record DeliveryWindowRow(
-    int Id,
+    long Id,
     string Label,
     TimeOnly StartsAt,
     TimeOnly EndsAt,
